refactoring and issue with powermock
