+++ app/build.gradle
@@ -384,7 +384,9 @@ dependencies {
     implementation "androidx.navigation:navigation-ui-ktx:$androidx_navigation_version"
     implementation 'net.openid:appauth:0.7.1'
     implementation "androidx.browser:browser:1.3.0"
-    implementation 'com.adamratzman:spotify-api-kotlin-android:3.4.03'
+    implementation ('com.adamratzman:spotify-api-kotlin-android:3.5.01') {
+        exclude group: 'org.jetbrains.kotlinx', module: 'kotlinx-coroutines-core'
+    }
 
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
