+++ build.gradle
@@ -2,7 +2,7 @@
 
 buildscript {
     ext.kotlin_version = '1.4.21'
-    ext.kotlin_coroutines_version = '1.4.2'
+    ext.kotlin_coroutines_version = '1.3.9'
     ext.androidx_lifecycle_extensions_version = '2.2.0'
     ext.androidx_navigation_version = '2.3.2'
     repositories {
